fix review hints, reduce code duplication
