+++ DcCore/DcCore/Extensions/String+Extensions.swift
@@ -1,7 +1,7 @@
 import Foundation
-internal extension String {
+public extension String {
     
-    static func localized(_ stringID: String) -> String {
+    public static func localized(_ stringID: String) -> String {
         let value = NSLocalizedString(stringID, comment: "")
         if value != stringID || NSLocale.preferredLanguages.first == "en" {
             return value
@@ -14,7 +14,7 @@ internal extension String {
         return NSLocalizedString(stringID, bundle: bundle, comment: "")
     }
 
-    static func localized(stringID: String, count: Int) -> String {
+    public static func localized(stringID: String, count: Int) -> String {
         let formatString: String = localized(stringID)
         let resultString: String = String.localizedStringWithFormat(formatString, count)
         return resultString
